Exceptions/Backend/requirements.txt
django
pillow
djangorestframework
django-rest-authtoken
django-crispy-forms
